revamp editor view
