+++ Valet/Views/License/LicenseInfo/LicenceInfo.swift
@@ -8,6 +8,7 @@ struct LicenceInfo: View {
 	@EnvironmentObject var formState: EditFormState
 	var license: License
 	
+	@State var showEditAppSheet: Bool = false
 	@State private var showToast: Bool = false
 	
 	@AppStorage("disableAnimations") private var disableAnimations: Bool = false
@@ -24,25 +25,18 @@ struct LicenceInfo: View {
 							.aspectRatio(contentMode: .fit)
 							.frame(width: 75)
 						VStack(alignment: .leading) {
-							if viewModes.editMode == true {
-								TextField("Some Cool App", text: $formState.softwareName)
-									.textFieldStyle(RoundedBorderTextFieldStyle())
-								TextField("https://sampleapp.com/download", text: $formState.urlString)
-									.textFieldStyle(RoundedBorderTextFieldStyle())
-							} else {
-								Text(license.softwareName)
-									.font(.title)
-									.multilineTextAlignment(.leading)
-								if let url = license.downloadUrl {
-									Link(destination: url, label: {
-										if isDownloadLink(url: url) {
-											Label("Download", systemImage: "arrow.down.circle")
-										} else {
-											Label("Website", systemImage: "safari")
-										}
-									})
-									.buttonStyle(.borderedProminent)
-								}
+							Text(license.softwareName)
+								.font(.title)
+								.multilineTextAlignment(.leading)
+							if let url = license.downloadUrl {
+								Link(destination: url, label: {
+									if isDownloadLink(url: url) {
+										Label("Download", systemImage: "arrow.down.circle")
+									} else {
+										Label("Website", systemImage: "safari")
+									}
+								})
+								.buttonStyle(.borderedProminent)
 							}
 						}
 						Spacer()
@@ -53,27 +47,23 @@ struct LicenceInfo: View {
 					LicenseInfoRow(
 						showToast: $showToast,
 						value: license.registeredToName,
-						formValue: $formState.registeredToName,
 						label: "Registered To"
 					)
 					
 					LicenseInfoRow(
 						showToast: $showToast,
 						value: license.registeredToEmail,
-						formValue: $formState.registeredToEmail,
 						label: "Email"
 					)
 					
 					LicenseInfoRow(
 						showToast: $showToast,
 						value: license.licenseKey,
-						formValue: $formState.licenseKey,
 						label: "License Key"
 					)
 					
 					DateInfoRow(
 						value: license.expirationDt,
-						formValue: $formState.expirationDt,
 						label: "Expiration Date"
 					)
 					
@@ -81,17 +71,11 @@ struct LicenceInfo: View {
 					Divider()
 					Text("Notes")
 						.font(.caption)
-					if viewModes.editMode == true {
-						TextEditor(text: $formState.notes)
-							.frame(minHeight: 100)
-					} else {
-						Markdown(license.notes)
-					}
+					Markdown(license.notes)
 				}
 				.frame(maxWidth: .infinity)
 				.padding()
 			}
-			.animation(disableAnimations == false ? .easeIn : nil, value: viewModes.editMode)
 		}
 		.frame(maxWidth: .infinity)
 		.environmentObject(formState)
@@ -102,30 +86,20 @@ struct LicenceInfo: View {
 			ToolbarItem {
 				Spacer()
 			}
-			if viewModes.editMode == true {
-				ToolbarItem {
-					Button(action: {
-						saveFormState()
-						viewModes.editMode.toggle()
-					}, label: {
-						Image(systemName: "checkmark.circle")
-					})
-					.disabled(!isEdited())
-					.keyboardShortcut(KeyEquivalent("s"))
-					.help("Save")
-				}
-			}
-			ToolbarItem {
+			ToolbarItem(placement: .primaryAction) {
 				Button(action: {
-					if viewModes.editMode == false {
+					if showEditAppSheet == false {
 						initFormState()
 					}
-					viewModes.editMode.toggle()
+					showEditAppSheet.toggle()
 				}, label: {
-					Image(systemName: viewModes.editMode == true ? "xmark.circle" : "square.and.pencil")
+					Image(systemName: "square.and.pencil")
 				})
-				.help(viewModes.editMode == true ? "Cancel" : "Edit")
+				.help("Edit")
 			}
+		}
+		.sheet(isPresented: $showEditAppSheet) {
+			EditAppView(isPresented: $showEditAppSheet, license: license)
 		}
 	}
 	
@@ -137,37 +111,5 @@ struct LicenceInfo: View {
 		formState.licenseKey = license.licenseKey
 		formState.expirationDt = license.expirationDt
 		formState.notes = license.notes
-	}
-	
-	private func saveFormState() {
-		do {
-			var updatedLicense = license
-			updatedLicense.softwareName = formState.softwareName
-			updatedLicense.downloadUrlString = formState.urlString
-			updatedLicense.registeredToName = formState.registeredToName
-			updatedLicense.registeredToEmail = formState.registeredToEmail
-			updatedLicense.licenseKey = formState.licenseKey
-			updatedLicense.expirationDt = formState.expirationDt
-			updatedLicense.notes = formState.notes
-			try updateLicense(databaseManager.dbQueue, data: updatedLicense)
-			databaseManager.fetchData()
-		} catch {
-			logger.error("ERROR: \(error)")
-		}
-	}
-	
-	private func isEdited() -> Bool {
-		if formState.softwareName == license.softwareName &&
-			formState.urlString == license.downloadUrlString &&
-			formState.registeredToName == license.registeredToName &&
-			formState.registeredToEmail == license.registeredToEmail &&
-			formState.licenseKey == license.licenseKey &&
-			formState.expirationDt == license.expirationDt &&
-			formState.notes == license.notes
-		{
-			return false
-		}
-		
-		return true
 	}
 }
